+++ packages/actions/test/setup-test-browser-env.ts
@@ -1,0 +1,4 @@
+import * as Turbo from '@hotwired/turbo';
+
+Turbo.session.drive = false;
+Turbo.start();