+++ packages/turbo-stream/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coldwired/turbo-stream
 
+## 0.9.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @coldwired/actions@0.9.0
+
 ## 0.8.1
 
 ### Patch Changes